Fix regression: Go to Definition stop working for variables inside command (#221)
